requirements.txt
faiss-cpu
sentence-transformers
numpy
openai
streamlit
scikit-learn
rouge-score
